Add the prebuilt rebar application and .cmd shim.
Now if I already knew chef well enough to get it to setup and environment for
me on windows and drop all these in the right path, I'd do that.  But alas,
I don't care that badly and I know that this repo isn't going to be used in
the future for anything critical.  So we're just going to check binaries into
git to make this easier and more reproducible.