Add a simple unit test to spin up the server.
This was to make sure that rebar could compile things right.
rebar compile should produce no errors.
rebar eunit should run the tests.
I didn't follow the -ifdef(TEST). trick that was mentioned in the rebar wiki
because I like having production code include tests.  It's always nice to be
able to run a self-test on a production environment to ensure that everything
is running as expected.