Add spec annotations.
I was hoping this would catch more type errors than it did.  It caught spelling
errors but it was rather...  inadequate at validating any type/data flow across
call backs such as gen_server or the supervisors or cowboy's rest handlers.  So
it was basically useless and a bit of a waste of time.  Oh well.  Now I know.